Split server.js into app.js udpate README
